Impresora class was printing with different formats. It violates single responsability principle.
First approach was to use an Abstract class and reimplement "imprimir" bahaviuor. But, in the end, that means inheritance and I'd prefer composition.

As we are only "decorating" the message, I choose "Decorator Pattern". ImpresoraHTML "decorates" ImpresoraSinFormato. In fact, it's only adding "html" tag to the message. We open a door for further decorators (json, xml...) by composition.

Our Main class is dirty, it smells badly. We have to do something with it