To remove all the logic from Main I've created a Factory Pattern.
It's not nice to have a factory called just once. With Abstract class and inheritance we could have implemented a cleaner solution, a Factory Method.

But, between extra-class and composition or inheritance, I do prefer the first solution. If we are not in Class Explosion danger (and we are not in this example).